feat: split postgres.go to multiple db implementation
